Ir sensor fixes

+++ src/shared_gui_delegate_on_robot.py
@@ -76,13 +76,13 @@ class DelegateThatReceives(object):
         self.robot.drive_system.go_straight_until_color_is_not(color, int(speed))
 
     def go_forward_until_distance_is_less_than(self, inches, speed):
-        self.robot.drive_system.go_forward_until_distance_is_less_than(int(inches), int(speed))
+        self.robot.drive_system.go_forward_until_distance_is_less_than(float(inches), int(speed))
 
     def go_forward_until_distance_is_greater_than(self, inches, speed):
-        self.robot.drive_system.go_backward_until_distance_is_greater_than(int(inches), int(speed))
+        self.robot.drive_system.go_backward_until_distance_is_greater_than(float(inches), int(speed))
 
     def go_until_distance_is_within(self, delta, inches, speed):
-        self.robot.drive_system.go_until_distance_is_within(int(delta), int(inches), int(speed))
+        self.robot.drive_system.go_until_distance_is_within(float(delta), float(inches), int(speed))
 
     # SoundSystem
     def beep_n_times(self, n):
